Refactor services: Added chain-data and genesis
