+++ src/v0/services/mod.rs
@@ -1,4 +1,5 @@
-mod ids;
+mod chain_data;
+mod genesis;
 use crate::v0::context::SharedContext;
 
 use warp::filters::BoxedFilter;
@@ -8,6 +9,9 @@ pub fn filter(
     root: BoxedFilter<()>,
     context: SharedContext,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
-    let ids_root = warp::path!("ids" / ..);
-    ids::filter(root.and(ids_root).boxed(), context)
+    let chain_data_root = warp::path!("chain-data" / ..);
+    let genesis_root = warp::path!("genesis" / ..);
+    let chain_data_filter = chain_data::filter(chain_data_root.boxed(), context.clone());
+    let genesis_filter = genesis::filter(genesis_root.boxed(), context.clone());
+    root.and(genesis_filter.or(chain_data_filter)).boxed()
 }
